wip #1500 Get image content for Richtext processing
